requirements.txt
langchain>=0.0.350
langchain-community>=0.0.10
langchain-openai>=0.0.2
boto3>=1.34.0
opensearch-py>=2.4.1
requests-aws4auth>=1.2.3
python-dotenv>=1.0.0 